fix: 32-bit-int sentinel + PPC signed-char (VNC message-parse crash)
The client connects, handshakes, sends ServerInit — then every client message
parses as all-zeros in a loop until a type-1 bus error. Two 16-bit-int-era
assumptions that only break on PowerPC (32-bit int, unsigned char):

- processMessageFragments resume sentinels are `unsigned short` set to 0xFFFF and
  checked with `!= -1`. On 32-bit int the short promotes to 65535 so `!= -1` is
  ALWAYS true: the deferred-message path re-enters forever, setPosition(65535)
  runs past the RDS into stale/returned buffers -> reads zeros -> crash. Compare
  against (unsigned short)-1 instead. (68k's 16-bit int made 0xFFFF == -1.)
- Build with -fsigned-char: the code was written for 68k (signed char); PowerPC's
  default unsigned char breaks `char` loop counters like `while (--n != -1)` in
  the encoders (the next crash after the connection one).

Diagnosed live via the UDP log sink (repeating "depth 0" SetPixelFormat + loop).

+++ mac-cpp-source/VNCServer.cpp
@@ -640,7 +640,12 @@ Boolean processMessageFragments (CallContext context) {
     static unsigned short resumeWriteAt = -1;
 
     if (context == asMainLoop) {
-        if (resumeReadAt != -1) {
+        // NB: resumeReadAt is unsigned short holding 0xFFFF as the "none" sentinel.
+        // Compare against (unsigned short)-1, not -1: with 32-bit int (PowerPC) the
+        // unsigned short promotes to 65535 and "!= -1" is ALWAYS true, so the
+        // deferred path re-enters forever, setPosition(65535) runs past the RDS and
+        // it reads zeros in a loop -> type-1 crash. (68k's 16-bit int hid this.)
+        if (resumeReadAt != (unsigned short)-1) {
             inStream.setPosition(resumeReadAt);
             resumeReadAt = -1;
             dprintf("\n==== Starting deferred messages ====\n");
@@ -652,7 +657,7 @@ Boolean processMessageFragments (CallContext context) {
 
     unsigned short msgStart = 0;
 
-    if (resumeWriteAt != -1) {
+    if (resumeWriteAt != (unsigned short)-1) {   // same 32-bit-int sentinel fix as above
         pb.msgAvail = resumeWriteAt;
         resumeWriteAt = -1;
         goto continueInterruptedMessage;
